# PetsClient/pets-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Pets API"
  },
  "paths": {
    "/api/Pets": {
      "get": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": { "$ref": "#/definitions/Pet" }
            }
          }
        }
      },
      "post": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsPost",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "pet",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Pet" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/Pets/{id}": {
      "get": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "put": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdPut",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pet",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Pet" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/Values": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": { "type": "string" }
            }
          }
        }
      },
      "post": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesPost",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/Values/{id}": {
      "get": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "type": "string" }
          }
        }
      },
      "put": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdPut",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "value",
            "in": "body",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "Values" ],
        "operationId": "ApiValuesByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    }
  },
  "definitions": {
    "Pet": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "owner": { "type": "string" },
        "name": { "type": "string" },
        "type": { "type": "string" },
        "age": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}